+++ examples/vue-app/src/views/AboutView.vue
@@ -2,6 +2,6 @@
 
 <template>
   <main>
-    <h1>About Page</h1>
+    <h2 class="rpl-type-h2">About Page</h2>
   </main>
 </template>